Reorg embedding code generator code
JIRA: RAIL-4069
